+++ src/query-builder/components/QueryBuilder.tsx
@@ -16,7 +16,8 @@ import ClauseList from './ClauseList';
 import useDataApi from '../../shared/hooks/useDataApi';
 
 import { createEmptyClause, defaultQueryFor } from '../utils/clause';
-import { stringify, parse } from '../utils/queryStringProcessor';
+import { stringify } from '../utils/queryStringProcessor';
+import parseAndMatchQuery from '../utils/parseAndMatchQuery';
 
 import { addMessage } from '../../messages/state/messagesActions';
 
@@ -32,101 +33,6 @@ import { Clause, SearchTermType } from '../types/searchTypes';
 
 import '../../uniprotkb/components/search/styles/search-container.scss';
 import './styles/query-builder.scss';
-
-type STTWithParent = SearchTermType & {
-  parent?: STTWithParent;
-};
-
-const flatten = (searchTermData: STTWithParent[]): STTWithParent[] => {
-  return searchTermData.flatMap((searchTermDatum: STTWithParent) => {
-    if (searchTermDatum.siblings) {
-      return flatten(searchTermDatum.siblings).map((st) => ({
-        ...st,
-        parent: searchTermDatum,
-      }));
-    }
-    if (searchTermDatum.items) {
-      return flatten(searchTermDatum.items);
-    }
-    return searchTermDatum;
-  });
-};
-
-const parseAndMatchQuery = (
-  query: string | string[] | null | undefined,
-  possibleSearchTerms: STTWithParent[]
-): [valid: Clause[], invalid: Clause[]] => {
-  const parsedQuery = query && !Array.isArray(query) ? parse(query) : undefined;
-  // for each parsed clause, try to find the corresponding endpoint-described
-  // clause to merge its 'searchTerm' field
-  const validatedQuery: Clause[] = [];
-  const invalid: Clause[] = [];
-  for (const clause of parsedQuery || []) {
-    if (clause.searchTerm.term === 'All') {
-      validatedQuery.push(clause);
-      continue; // eslint-disable-line no-continue
-    }
-    const matching = possibleSearchTerms.filter(
-      ({ term }) => term === clause.searchTerm.term
-    );
-    // if it exists, assign it 'searchTerm'
-    if (matching.length) {
-      if (matching.length === 1) {
-        // only one search term matching
-        validatedQuery.push({
-          ...clause,
-          searchTerm: matching[0].parent || matching[0],
-        });
-      } else if (clause.searchTerm.term === 'xref') {
-        // specific case for crosss-references
-        const [prefix, ...rest] = clause.queryBits.xref.split('-');
-        const matchingXref = matching.find(
-          ({ valuePrefix }) => valuePrefix === `${prefix}-`
-        );
-        if (matchingXref) {
-          validatedQuery.push({
-            ...clause,
-            searchTerm: matchingXref,
-            queryBits: { xref: rest.join('-') },
-          });
-        } else {
-          // invalid xref prefix
-          invalid.push(clause);
-        }
-      } else {
-        invalid.push(clause);
-      }
-      // else, didn't find any match
-    } else {
-      // try to find a search term matching through the autoComplete field
-      const matchingAutoComplete = possibleSearchTerms.find(
-        ({ autoCompleteQueryTerm }) =>
-          autoCompleteQueryTerm &&
-          autoCompleteQueryTerm === clause.searchTerm.term
-      );
-      if (matchingAutoComplete) {
-        // set that clause to the corresponding autoComplete term field
-        validatedQuery.push({
-          ...clause,
-          searchTerm: matchingAutoComplete,
-          // and change queryBit key
-          queryBits: {
-            ...clause.queryBits,
-            [matchingAutoComplete.autoCompleteQueryTerm &&
-            clause.searchTerm.term.endsWith('_id')
-              ? matchingAutoComplete.autoCompleteQueryTerm
-              : matchingAutoComplete.term]: clause.queryBits[
-              clause.searchTerm.term
-            ],
-          },
-        });
-      } else {
-        invalid.push(clause);
-      }
-    }
-  }
-  return [validatedQuery, invalid];
-};
 
 const QueryBuilder: FC = () => {
   const history = useHistory();
@@ -173,14 +79,11 @@ const QueryBuilder: FC = () => {
         return clauses;
       }
 
-      // flatten all the endpoint-described clauses to be able to to look-up
-      const flattened = flatten(searchTermsData);
-
       const query = qs.parse(location.search, { decode: true })?.query;
 
       const [validatedQuery, invalidClauses] = parseAndMatchQuery(
         query,
-        flattened
+        searchTermsData
       );
 
       if (invalidClauses.length) {
@@ -205,7 +108,7 @@ const QueryBuilder: FC = () => {
         return validatedQuery;
       }
 
-      return parseAndMatchQuery(defaultQueryFor(namespace), flattened)[0];
+      return parseAndMatchQuery(defaultQueryFor(namespace), searchTermsData)[0];
     });
   }, [dispatch, location.search, loading, namespace, searchTermsData]);
